Added main for date
Pretty straight forward, it's the best example of what's currently available in terms of features when it comes to the cactus for the corresponding release data and version. Although, when it comes to it, the subroutines would have needed a little more work for this version to succesfully run roots. So don't try it